Iron 0.21 — wasm-gc executor: extract into aver::runtime::wasm_gc
Pulls the compile-and-run pipeline out of the `aver` binary so
embedders (parity fuzz target, verify executor, future REPL) can
share the same path the `aver run --wasm-gc` CLI takes — without
spawning a subprocess or going through `process::exit`.

Layout:

  src/runtime/mod.rs               new module root
  src/runtime/wasm_gc.rs           moved from src/main/run_wasm_gc.rs
  src/runtime/wasm_gc/decode.rs    moved
  src/runtime/wasm_gc/imports.rs   moved (+ imports/* submodules)

`EffectMode::Recording(&str dir)` and the on-disk trace persistence
go away from the lib path; the executor now hands the captured
trace back in `RunOutcome::recorded_effects` and the CLI wrapper
turns it into a `SessionRecording` + writes the JSON. File I/O
is a CLI concern, not the runtime's. `EffectMode::Replaying` →
`EffectMode::Replay` along the way (shorter name, lib-clean).

src/main/run_wasm_gc.rs shrinks to ~230 LOC of CLI glue:

  - `cmd_run_wasm_gc` — pretty-print + exit wrapper
  - `cmd_run_wasm_gc_with_mode` — same, with explicit mode
  - `try_run_wasm_gc` — file read + multi-module flatten + lib
    call + recording persistence; the only function `replay_cmd`
    consumes directly

The lib entry point is:

    aver::runtime::wasm_gc::run_in_process(
        items, analysis,
        RunConfig { program_args, entry_info, mode },
    ) -> Result<RunOutcome, String>

`run_in_process` takes already-parsed `&[TopLevel]` so the fuzz
target can feed it the same items it just typechecked, without
paying the read-from-disk + reparse cost per AFL iteration.

627 lib tests pass; the `cargo build --features wasm` binary is
byte-identical to before the move (only file paths changed).
