+++ src/runtime/wasm_gc/imports/console.rs
