+++ src/runtime/wasm_gc/imports/disk.rs
